Update ignore, include compiled assets
I used to deploy app files only, ignoring compiled assets. Now, we'll
compile locally and send a clean assets folder to the repo for deployment.

This makes testing easier and cuts down on issues with assets being overridden
by old bundles and removes differences in app display when in production.